Improve condition

+++ homeassistant/components/shelly/button.py
@@ -214,7 +214,10 @@ async def async_setup_entry(
 
     # Remove the 'restart' button for sleeping devices as it was mistakenly
     # added in https://github.com/home-assistant/core/pull/154673
-    if coordinator.sleep_period > 0:
+    if (
+        get_device_entry_gen(config_entry) in RPC_GENERATIONS
+        and coordinator.sleep_period > 0
+    ):
         async_remove_shelly_entity(hass, BUTTON_PLATFORM, f"{coordinator.mac}-reboot")
 
     entities: list[ShellyButton] = []
